requirements.txt
Django==5.1.3
django-colorfield==0.11.0
djangorestframework==3.15.2
